feat: 子分组批量转为一级分组 + pluginId 变更一键更新引用
问题4: 含子分组的一级分组菜单加「转为一级分组」多选对话框，批量把选中子分组提升为独立一级分组（新建同名分组、配置项移入、categoryPath 清空、子分组音频参数迁移到新分组 audioParams、源分组同步清理）。

问题3: PluginManagerScreen 内联编辑保存后，若 pluginId 变更则检测引用旧 id 的 TTS 配置项，弹窗提示并支持一键更新为新 id。ToolBoxScreen/PluginTtsUI 对「角色管理」插件按 id 或名称回退识别。

附带编译修复: App.kt 缺 SystemTtsConfig import、GroupItem.kt 的 AccountTree 改用 Icons.Default、ListManagerScreen 的 awaitFrame 改 withFrameNanos、LazyListScope 内移除 remember 调用。

+++ lib-tts/src/main/java/com/github/jing332/tts/DefaultResultProcessor.kt
@@ -162,7 +162,7 @@ internal class DefaultResultProcessor(
                     if (pcm != null) {
                         if (loudnessPcmCollector != null) {
                             val dup = pcm.duplicate()
-                            while (dup.hasRemaining()) loudnessPcmCollector.write(dup.get())
+                            while (dup.hasRemaining()) loudnessPcmCollector.write(dup.get().toInt())
                         }
                         callback.receive(pcm)
                     }
@@ -174,7 +174,7 @@ internal class DefaultResultProcessor(
                     val out = processor.output
                     if (loudnessPcmCollector != null && out.hasRemaining()) {
                         val dup = out.duplicate()
-                        while (dup.hasRemaining()) loudnessPcmCollector.write(dup.get())
+                        while (dup.hasRemaining()) loudnessPcmCollector.write(dup.get().toInt())
                     }
                     callback.receive(out)
                 } else {
@@ -183,7 +183,7 @@ internal class DefaultResultProcessor(
                         val out = processor.output
                         if (loudnessPcmCollector != null && out.hasRemaining()) {
                             val dup = out.duplicate()
-                            while (dup.hasRemaining()) loudnessPcmCollector.write(dup.get())
+                            while (dup.hasRemaining()) loudnessPcmCollector.write(dup.get().toInt())
                         }
                         callback.receive(out)
                     }
